+++ src/Console/Commands/Scaffold/FilterCommand.php
@@ -198,7 +198,7 @@ class FilterCommand extends BaseCommand
         $parts = explode('/', str_replace('\\', '/', $name));
 
         foreach ($parts as $part) {
-            if (!preg_match('/^[A-Z][a-zA-Z0-9]*$/', $part)) {
+            if (preg_match('/^[A-Z][a-zA-Z0-9]*$/', $part) !== 1) {
                 return false;
             }
         }