+++ src/Console/Commands/Scaffold/RuleCommand.php
@@ -174,7 +174,7 @@ class RuleCommand extends BaseCommand
         $parts = explode('/', str_replace('\\', '/', $name));
 
         foreach ($parts as $part) {
-            if (!preg_match('/^[A-Z][a-zA-Z0-9]*$/', $part)) {
+            if (preg_match('/^[A-Z][a-zA-Z0-9]*$/', $part) !== 1) {
                 return false;
             }
         }